Worked through checkbox challenge...
This was a small challenge in which I started becoming more familiar with new React 19 features.  Here, we have a small set of checkboxes in which I had to add 'name' and 'value' attributes so that they could be sent to my formData.  The 'getAll' method can be conveniently used in conjunction with 'formData' to grab multiple checkbox elements simultaneously, as seen here in the code (lines 8 and 9).