Create nightAtTheScrimbies
This was a JavaScript challenge in which I had to create a mess of nested arrays, and flatten out all of the elements in the arrays in an effort to give random awards to the listed nominees.  We completed this by mapping through the data, grabbing the hosts arrays, and concatenating them together into a single array.  After that, we created a random number generator for the index of the awards array, and used template literals to present the new data.  On each map iteration, a random number is generated for each element in the array, assigning a different number to each person.